+++ rules/nette-kdyby/tests/Rector/ClassMethod/ReplaceMagicPropertyWithEventClassRector/ReplaceMagicPropertyWithEventClassRectorTest.php
