+++ rules/symfony3/src/Rector/MethodCall/ChangeStringCollectionOptionToConstantRector.php
@@ -15,9 +15,9 @@ use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
  * @see https://symfony.com/doc/3.0/form/form_collections.html
  * @see https://symfony2-document.readthedocs.io/en/latest/reference/forms/types/collection.html#type
  *
- * @see \Rector\Symfony3\Tests\Rector\MethodCall\ChangeCollectionTypeOptionTypeFromStringToClassReferenceRector\ChangeCollectionTypeOptionTypeFromStringToClassReferenceRectorTest
+ * @see \Rector\Symfony3\Tests\Rector\MethodCall\ChangeStringCollectionOptionToConstantRector\ChangeStringCollectionOptionToConstantRectorTest
  */
-final class ChangeCollectionTypeOptionTypeFromStringToClassReferenceRector extends AbstractFormAddRector
+final class ChangeStringCollectionOptionToConstantRector extends AbstractFormAddRector
 {
     public function getRuleDefinition(): RuleDefinition
     {
